Move styling to separate file in component folder
