+++ src/components/admin/GuildList/style.js
@@ -1,0 +1,7 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles((theme) => ({
+  paper: { width: 800, background: theme.palette.background.secondary },
+}));
+
+export default useStyles;
