+++ src/components/auth/Logout/style.js
@@ -1,0 +1,11 @@
+import { makeStyles } from "@material-ui/core";
+
+const SIDEBAR_WIDTH = 200;
+
+const useStyles = makeStyles(() => ({
+  backdrop: {
+    left: SIDEBAR_WIDTH,
+  },
+}));
+
+export default useStyles;
