+++ src/components/intel/CampaignSelector/style.js
@@ -1,0 +1,12 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles(() => ({
+  select: { minWidth: 200 },
+  button: {
+    marginRight: -25,
+    marginLeft: -25,
+    transform: "translate(40%, 15%)",
+  },
+}));
+
+export default useStyles;
