+++ src/components/intel/NewWorld/style.js
@@ -1,0 +1,10 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles(() => ({
+  map: {
+    width: "100%",
+    height: "90vh",
+  },
+}));
+
+export default useStyles;
