+++ src/components/intel/PinDetails/style.js
@@ -1,0 +1,5 @@
+import { Typography, withStyles } from "@material-ui/core";
+
+export const StyledTypography = withStyles({ root: { margin: -5 } })(
+  Typography
+);
