+++ src/components/intel/World/style.js
@@ -1,0 +1,10 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles(() => ({
+  map: {
+    width: "90%",
+    height: "75vh",
+  },
+}));
+
+export default useStyles;
