+++ src/components/main/Main/style.js
@@ -1,0 +1,11 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles((theme) => ({
+  content: {
+    flexGrow: 1,
+    passing: theme.spacing(3),
+    marginTop: 20,
+  },
+}));
+
+export default useStyles;
