+++ src/components/utilities/Upload/style.js
@@ -1,0 +1,9 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles({
+  input: {
+    display: "none",
+  },
+});
+
+export default useStyles;
